refactor: reorganize client functions into services and update imports in forms
